added doxygen comments for utils
