+++ src/utils/env_utils.c
@@ -1,6 +1,18 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   env_utils.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: cwenz <cwenz@student.42.fr>                +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/25 17:11:13 by cwenz             #+#    #+#             */
+/*   Updated: 2023/11/25 17:22:13 by cwenz            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "minishell.h"
 
+/// @brief Adds a new node to the env variable linked list
 int	add_env_node(t_env **head, char *key, char *full_string, char *value)
 {
 	t_env	*temp;
@@ -13,7 +25,6 @@ int	add_env_node(t_env **head, char *key, char *full_string, char *value)
 	new_node->full_string = full_string;
 	new_node->value = value;
 	new_node->next = NULL;
-
 	if (!(*head))
 		*head = new_node;
 	else
@@ -26,12 +37,10 @@ int	add_env_node(t_env **head, char *key, char *full_string, char *value)
 	return (EXIT_SUCCESS);
 }
 
-/**
- * @brief Finds the node which holds the given key.
- * @param head A pointer to the first head of the env array
- * @param key The key to search for.
- * @return The node which holds the given key, else NULL if not found
- */
+/// @brief Finds the node which holds the given key.
+/// @param head A pointer to the first head of the env array
+/// @param key The key to search for.
+/// @return The node which holds the given key, else NULL if not found
 t_env	*find_env_key(t_env *head, char *key)
 {
 	t_env	*temp;
@@ -46,6 +55,9 @@ t_env	*find_env_key(t_env *head, char *key)
 	return (NULL);
 }
 
+/// @brief Copes the given `envp` into a linked list format
+/// @param head A pointer to which the data will be added
+/// @param envp The array which will be used to create the linked list.
 void	copy_env_variables(t_env **head, char **envp)
 {
 	int	i;
