+++ src/utils/utils.c
@@ -1,3 +1,14 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   utils.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: cwenz <cwenz@student.42.fr>                +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/25 17:16:21 by cwenz             #+#    #+#             */
+/*   Updated: 2023/11/25 17:23:26 by cwenz            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "minishell.h"
 
@@ -24,18 +35,9 @@ char	*strjoin_pipex(char *s1, char *s2)
 	return (str);
 }
 
-void	copy_std_fds(t_minishell *minishell)
-{
-	minishell->stdfds.stdin = dup(STDIN_FILENO);
-	minishell->stdfds.stdout = dup(STDOUT_FILENO);
-}
-
-void	reset_fds(t_minishell *minishell)
-{
-	dup2(minishell->stdfds.stdin, STDIN_FILENO);
-	dup2(minishell->stdfds.stdout, STDOUT_FILENO);
-}
-
+/// @brief Adds the given pid to the linked list
+/// @param minishell The struct which holds the shell info
+/// @param pid The pid to add to the linked list
 void	add_pid(t_minishell *minishell, int pid)
 {
 	t_pids	*temp;
@@ -56,6 +58,10 @@ void	add_pid(t_minishell *minishell, int pid)
 	}
 }
 
+/// @brief Finds the given pid in the linked list
+/// @param minishell The struct containing the pids
+/// @param pid The pid to find
+/// @return The pid (if found), otherwise NULL
 t_pids	*find_pid(t_minishell *minishell, int pid)
 {
 	t_pids	*temp;
@@ -70,6 +76,8 @@ t_pids	*find_pid(t_minishell *minishell, int pid)
 	return (NULL);
 }
 
+/// @brief Loops through the linked list of pids and frees it
+/// @param minishell The struct containing the linked list fo pids
 void	free_pids(t_minishell *minishell)
 {
 	t_pids	*temp;
